fix(tests): revise old parser tests and relocate properly
